refactor(api-service): extract adapter postable message builder
applyOutboundBranding and buildAdapterPostableMessage were pure
functions living as private gateway methods, forcing the spec to
construct the Nest service with stubbed deps and reach in via `as any`.
Move them to an egress sibling module, test them directly, and trim the
web chat e2e assertion to the single field branding would mutate.
